Revert "GH-15: Refactor file_manager / FileUtils (#20)"
This reverts commit f4b652f6562c4997e4aaf679e5929fb51f0bae9a.
